Moved Parameter and Tunable to com.sap.sgs.phosphor.fosstars.model
This is a part of #40
